# static/api/openapi.json
{
  "openapi": "3.0.0",
  "info": {
    "title": "Idling.app API",
    "version": "0.338.7",
    "description": "Complete API documentation for Idling.app - a modern social platform",
    "contact": {
      "name": "Idling.app Development Team",
      "url": "https://github.com/Underwood-Inc/idling.app__UI"
    },
    "license": {
      "name": "MIT",
      "url": "https://opensource.org/licenses/MIT"
    }
  },
  "servers": [
    {
      "url": "http://localhost:3000",
      "description": "Development server (Next.js)"
    },
    {
      "url": "https://idling.app",
      "description": "Production server"
    }
  ],
  "components": {
    "securitySchemes": {
      "NextAuth": {
        "type": "apiKey",
        "in": "cookie",
        "name": "next-auth.session-token",
        "description": "NextAuth session cookie"
      },
      "AdminRole": {
        "type": "apiKey",
        "in": "header",
        "name": "X-Admin-Role",
        "description": "Admin role verification"
      }
    },
    "schemas": {
      "Error": {
        "type": "object",
        "properties": {
          "error": {
            "type": "string",
            "description": "Error message"
          },
          "details": {
            "type": "string",
            "description": "Additional error details"
          }
        },
        "required": [
          "error"
        ]
      },
...
      "EmojiResponse": {
        "type": "object",
        "properties": {
          "id": {
            "type": "number",
            "description": "Emoji ID"
          },
          "emoji_id": {
            "type": "string",
            "description": "Emoji identifier"
          },
          "unicode_codepoint": {
            "type": "string",
            "description": "Unicode codepoint"
          },
          "unicode_char": {
            "type": "string",
            "description": "Unicode character"
          },
          "name": {
            "type": "string",
            "description": "Emoji name"
          },
          "description": {
            "type": "string",
            "description": "Emoji description"
          },
          "category": {
            "type": "object",
            "properties": {
              "id": {
                "type": "number"
              },
              "name": {
                "type": "string"
              },
              "display_name": {
                "type": "string"
              }
            }
          },
          "tags": {
            "type": "array",
            "items": {
              "type": "string"
            },
            "description": "Emoji tags"
          },
          "aliases": {
            "type": "array",
            "items": {
              "type": "string"
            },
            "description": "Emoji aliases"
          },
          "keywords": {
            "type": "array",
            "items": {
              "type": "string"
            },
            "description": "Emoji keywords"
          },
          "is_custom": {
            "type": "boolean",
            "description": "Whether this is a custom emoji"
          },
          "custom_image_url": {
            "type": "string",
            "description": "URL for custom emoji image"
          },
          "usage_count": {
            "type": "number",
            "description": "Usage count for this emoji"
          },
          "version_min": {
            "type": "string",
            "description": "Minimum OS version required"
          },
          "is_approved": {
            "type": "boolean",
            "description": "Whether custom emoji is approved"
          },
          "is_personal": {
            "type": "boolean",
            "description": "Whether emoji is personal/private"
          }
        }
      },
      "EmojiListResponse": {
        "type": "object",
        "properties": {
          "emojis": {
            "type": "array",
            "items": {
              "$ref": "#/components/schemas/EmojiResponse"
            },
            "description": "List of emojis"
          },
          "categories": {
            "type": "array",
            "items": {
              "type": "object",
              "properties": {
                "id": {
                  "type": "number"
                },
                "name": {
                  "type": "string"
                },
                "display_name": {
                  "type": "string"
                },
                "description": {
                  "type": "string"
                },
                "emoji_count": {
                  "type": "number"
                }
              }
            },
            "description": "Available emoji categories"
          },
          "os_info": {
            "type": "object",
            "properties": {
              "os": {
                "type": "string"
              },
              "version": {
                "type": "string"
              },
              "is_supported": {
                "type": "boolean"
              },
              "emoji_support": {
                "type": "object",
                "properties": {
                  "supports_unicode": {
                    "type": "boolean"
                  },
                  "supports_custom": {
                    "type": "boolean"
                  },
                  "max_emoji_version": {
                    "type": "string"
                  },
                  "recommended_format": {
                    "type": "string"
                  }
                }
              }
            },
            "description": "Operating system information"
          },
          "total_count": {
            "type": "number",
            "description": "Total number of emojis"
          },
          "page": {
            "type": "number",
            "description": "Current page number"
          },
          "per_page": {
            "type": "number",
            "description": "Number of emojis per page"
          }
        }
      },
...
    }
  },
...
  "paths": {
...
    "/api/emojis": {
      "get": {
        "summary": "Get emojis by OS and category",
        "description": "Fetches OS-specific emojis with optional filtering by category and search terms",
        "tags": [
          "Emojis"
        ],
        "parameters": [
          {
            "in": "query",
            "name": "category",
            "schema": {
              "type": "string"
            },
            "description": "Filter by emoji category (e.g., 'smileys', 'animals')"
          },
          {
            "in": "query",
            "name": "search",
            "schema": {
              "type": "string"
            },
            "description": "Search term to filter emojis by name"
          },
          {
            "in": "query",
            "name": "page",
            "schema": {
              "type": "integer",
              "default": 1
            },
            "description": "Page number for pagination"
          },
          {
            "in": "query",
            "name": "per_page",
            "schema": {
              "type": "integer",
              "default": 50,
              "maximum": 100
            },
            "description": "Number of emojis per page"
          },
          {
            "in": "query",
            "name": "include_custom",
            "schema": {
              "type": "boolean",
              "default": false
            },
            "description": "Include custom user-uploaded emojis"
          }
        ],
        "responses": {
          "200": {
            "description": "Successfully retrieved emojis",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/EmojiListResponse"
                }
              }
            }
          },
          "400": {
            "description": "Invalid request parameters",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Error"
                }
              }
            }
          },
          "500": {
            "description": "Internal server error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Error"
                }
              }
            }
          }
        }
      },
      "post": {
        "summary": "Upload custom emoji",
        "description": "Upload a custom emoji image for the authenticated user",
        "tags": [
          "Emojis"
        ],
        "security": [
          {
            "NextAuth": []
          }
        ],
        "requestBody": {
          "required": true,
          "content": {
            "multipart/form-data": {
              "schema": {
                "type": "object",
                "properties": {
                  "file": {
                    "type": "string",
                    "format": "binary",
                    "description": "Emoji image file (PNG, JPG, GIF)"
                  },
                  "name": {
                    "type": "string",
                    "description": "Name for the custom emoji"
                  },
                  "category": {
                    "type": "string",
                    "description": "Category for the emoji"
                  }
                },
                "required": [
                  "file",
                  "name"
                ]
              }
            }
          }
        },
        "responses": {
          "201": {
            "description": "Emoji uploaded successfully",
            "content": {
              "application/json": {
                "schema": {
                  "type": "object",
                  "properties": {
                    "success": {
                      "type": "boolean"
                    },
                    "emoji": {
                      "$ref": "#/components/schemas/EmojiResponse"
                    }
                  }
                }
              }
            }
          },
          "400": {
            "description": "Invalid file or parameters"
          },
          "401": {
            "description": "Authentication required"
          },
          "413": {
            "description": "File too large"
          },
          "500": {
            "description": "Upload failed"
          }
        }
      },
      "put": {
        "summary": "Track emoji usage",
        "description": "Track usage statistics for an emoji",
        "tags": [
          "Emojis"
        ],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "type": "object",
                "properties": {
                  "emoji_id": {
                    "type": "string",
                    "description": "ID of the emoji being used"
                  },
                  "emoji_type": {
                    "type": "string",
                    "enum": [
                      "windows",
                      "mac",
                      "custom"
                    ],
                    "description": "Type of emoji being tracked"
                  }
                },
                "required": [
                  "emoji_id",
                  "emoji_type"
                ]
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "Usage tracked successfully",
            "content": {
              "application/json": {
                "schema": {
                  "type": "object",
                  "properties": {
                    "success": {
                      "type": "boolean"
                    }
                  }
                }
              }
            }
          },
          "400": {
            "description": "Invalid parameters"
          },
          "500": {
            "description": "Tracking failed"
          }
        }
      }
    },
...
  }
}